GROK-13354: Bio | Tools: Rearrange top menu, Dendrogram
Bio top menu rearranged
Fix tests
