+++ packages/Dendrogram/package.json
@@ -5,14 +5,14 @@
     "name": "Aleksandr Tanas",
     "email": "atanas@datagrok.ai"
   },
-  "version": "1.2.17",
+  "version": "1.2.18",
   "description": "Computation and visualization of hierarchical data clustering. [Learn more](https://github.com/datagrok-ai/public/blob/master/packages/Dendrogram/README.md)",
   "dependencies": {
-    "@datagrok-libraries/bio": "^5.32.1",
+    "@datagrok-libraries/bio": "^5.33.2",
     "@datagrok-libraries/gridext": "^1.3.56",
     "@datagrok-libraries/math": "^1.0.2",
     "@datagrok-libraries/ml": "^6.3.31",
-    "@datagrok-libraries/utils": "^4.0.7",
+    "@datagrok-libraries/utils": "^4.0.17",
     "cash-dom": "latest",
     "datagrok-api": "^1.15.0",
     "dayjs": "latest",
