+++ packages/Dendrogram/src/package-test.ts
@@ -12,6 +12,7 @@ import './tests/package-funcs-tests';
 import './tests/distance-matrix-tests';
 import './tests/calc-matrix-tests';
 import './tests/cluster-matrix-tests';
+import './tests/viewers';
 
 export const _package = new DG.Package();
 export {tests};
