+++ packages/Dendrogram/src/tests/grid-with-tree-tests.ts
@@ -2,17 +2,19 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
-import {category, test, testViewer} from '@datagrok-libraries/utils/src/test';
-import {injectTreeForGridUI2} from '../viewers/inject-tree-for-grid2';
-import {TreeHelper} from '../utils/tree-helper';
-import {_package} from '../package-test';
-import {viewsTests} from './utils/views-tests';
+import $ from 'cash-dom';
+
+import {awaitCheck, category, test, testViewer} from '@datagrok-libraries/utils/src/test';
 import {ITreeHelper} from '@datagrok-libraries/bio/src/trees/tree-helper';
 import {NodeType} from '@datagrok-libraries/bio/src/trees';
 import {parseNewick} from '@datagrok-libraries/bio/src/trees/phylocanvas';
 
+import {injectTreeForGridUI2} from '../viewers/inject-tree-for-grid2';
+import {TreeHelper} from '../utils/tree-helper';
 
-category('GridWithTree', viewsTests((ctx: { dfList: DG.DataFrame[], vList: DG.ViewBase[] }) => {
+import {_package} from '../package-test';
+
+category('GridWithTree', () => {
   test('open', async () => {
     const _th: ITreeHelper = new TreeHelper();
 
@@ -24,24 +26,13 @@ category('GridWithTree', viewsTests((ctx: { dfList: DG.DataFrame[], vList: DG.Vi
     const newickRoot: NodeType = parseNewick(newickStr);
 
     const tv: DG.TableView = grok.shell.addTableView(dataDf);
-    ctx.dfList.push(dataDf);
-    ctx.vList.push(tv);
-    const neighborWidth = 250;
-
+    await awaitCheck(() => {
+      return $(tv.root).find('.d4-grid canvas').length > 0;
+    }, 'The view grid canvas not found', 200);
+    ;const neighborWidth = 250;
     injectTreeForGridUI2(tv.grid, newickRoot, leafColName, neighborWidth);
+    await awaitCheck(() => {
+      return $(tv.root).find('.ui-div canvas').length == 1;
+    }, 'Injected tree not found', 200);
   });
-}));
-
-
-category('viewers', () => {
-  const viewers = DG.Func.find({package: 'Dendrogram', tags: ['viewer']}).map((f) => f.friendlyName);
-  for (const v of viewers) {
-    test(v, async () => {
-      await testViewer(v, await (async () => {
-        const newickStr: string = await _package.files.readAsText('data/tree95.nwk');
-        const treeHelper = new TreeHelper();
-        return treeHelper.newickToDf(newickStr, 'tree95');
-      })(), {detectSemanticTypes: true});
-    });
-  }
 });
